added dynamic descriptor/layout depending on shaders
