+++ Better-than-Nothing/Modules/Core/src/Renderer/Descriptor.hpp
@@ -1,0 +1,19 @@
+#pragma once
+
+namespace BetterThanNothing
+{
+	class Device;
+
+	class Descriptor
+	{
+	private:
+		Device* m_Device;
+
+		VkDescriptorSet m_DescriptorSets;
+
+	public:
+		Descriptor(Device* device, VkDescriptorSet descriptorSet);
+
+		void Update(u32 binding, VkDescriptorType type, VkDescriptorBufferInfo* bufferInfo, VkDescriptorImageInfo* imageInfo);
+	};
+};
